BUGFIX: Use different seed in second episode.
Before this fix the second episode used the same seed as the first episode (but subsequent episodes used different seeds).

- Ensure we use env_seed + 1 on the second episode.
- Extend tests of determinism.

NOTE: This did not affect the experiments in the Melting Pot paper as these were run with an older (internal) version of this code. The bug was introduced during the open-sourcing process.
PiperOrigin-RevId: 400227478
Change-Id: I57cb1c220e507c05c15e38acfc0a1b7e8725cb4b