BUGFIX: Use correct is_focal in team games (CTF and KotH).
Affected scenarios:
-   `capture_the_flag_0`
-   `capture_the_flag_1`
-   `king_of_the_hill_0`
-   `king_of_the_hill_1`
-   `king_of_the_hill_2`

Previously, we set `is_focal = (True,) * 4 + (False,) * 4` for team-vs-team scenarios. This mistakenly lead to focal players filling 2 player slots on one team and 2 player slots on the other. Since CTF is a zero-sum game this mean that the focal per-capita return was guaranteed to be zero.

Here, we set `is_focal = (True, False,) * 4`. This ensures that a team of 4 focal players plays against a team of 4 background bots, as intended.

(NOTE: this didn't affect our experiments run for the Melting Pot paper since these weren't run using Scenario.)

PiperOrigin-RevId: 432223638
Change-Id: I9c001183590db4464ec2c7ab1cb80840828374a7