Fix Lua bugs and upgrade macos build to Lua 5.2
We were relying on the order of components in the game objects' components table to be iterable in the same order as they were inserted. This is not guaranteed. We now make explicit the initialisation (via `start`) of the StateManager and the Transform (in that order) before any other component's `start` function is called.

fixes https://github.com/deepmind/meltingpot/issues/84
fixes https://github.com/deepmind/meltingpot/issues/24.

PiperOrigin-RevId: 492464124
Change-Id: I0ef2958cd4ca093a18d54a38a6b044b1265c4fc1